Rough Draft #1
Point 1: Summarize the two mindsets and characteristics of each
Growth Mindset is an idea that Professor Carol Dweck came up with while researching the way in which people learn. It was through this research that she developed the idea of Growth Mindset and how it differs from the more traditionally accepted way of learning which she calls the Fixed Mindset. These mindsets are attitudes towards learning and the beliefs about the roots of success. In her research and when she gives interviews/public talks she lists out the characteristics of each.
She describes Fixed Mindset as “I just have a certain amount of intelligence and that’s it”. It is a mindset that attaches labels and limits to one’s learning ability. A person with this mindset might display some of these characteristics; run from failure, little brain activity when problem solving, and a “I am never going to succeed” attitude. Growth Mindset on the other hand is the belief that intelligence can be developed. That success comes from effort, practice and hard work. Characteristics of someone with this mindset are; an understanding that abilities can be developed, higher brain activity in problem solving, high engagement and a “I have not succeed…..yet” attitude.
Point 2: Describe which aspects of each mindset I recognize in myself
For myself I recognised aspects of both these mindsets while I was conducting my research. I’ve had difficulties in the past overcoming labels that were placed on me either by others or myself. So in academia when I didn’t live up to those labels I felt like a failure and would in a way place new “negative” labels on myself. That was certainly damaging to my attempts at learning. Without realizing it I have been adopting aspects of Growth Mindset. Specifically the idea that it’s much easier to learn when I do not stop or get discouraged at the first sign of failure but rather keep pushing forward knowing that with hard work and practice the desired results will follow.
Point 3: Explain how I will apply this understanding to my experience here at The Iron Yard
On Day 2 and 3 of my Iron Yard experience I already got the chance to apply this understanding. And I did not even know it until I sat down to write this essay! I was incredibly discouraged  at my inability to learn and complete the homework assignments. So much so that I was questioning if I should withdraw from the program. I had to realize that these “road blocks” were not fatal but in fact just stepping stones towards my eventually mastering of coding. That by continuing to work hard I would succeed. I plan on reminding myself of this the next time I think I have failed.